+++ packages/combobox/src/Combobox.tsx
@@ -593,6 +593,7 @@ export function Combobox<M extends boolean>({
     }
   }, [inputValue, isOpen, prevValue, updateHighlightedOption]);
 
+  // TODO: Replace this with hooks/useAutoScroll
   // When the focused option changes, update the menu scroll if necessary
   useEffect(() => {
     if (highlightedOption) {
@@ -623,7 +624,7 @@ export function Combobox<M extends boolean>({
         const { value, displayName } = getNameAndValue(child.props);
 
         if (shouldOptionBeVisible(value)) {
-          const { className, glyph, disabled } = child.props;
+          const { className, glyph, disabled, ...rest } = child.props;
           const index = allOptions.findIndex(opt => opt.value === value);
 
           const isFocused = highlightedOption === value;
@@ -645,6 +646,7 @@ export function Combobox<M extends boolean>({
 
           return (
             <InternalComboboxOption
+              {...rest}
               value={value}
               displayName={displayName}
               isFocused={isFocused}
@@ -754,12 +756,12 @@ export function Combobox<M extends boolean>({
       e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
     ) => {
       if (!disabled) {
+        // Prevents triggering the setOpen function called by clicking anywhere within the input wrapper.
+        e.stopPropagation();
         updateSelection(null);
         onClear?.(e);
         onFilter?.('');
-        if (!isOpen) {
-          openMenu();
-        }
+        setInputFocus();
       }
     };
 
@@ -801,6 +803,7 @@ export function Combobox<M extends boolean>({
     onClear,
     onFilter,
     isOpen,
+    setInputFocus,
   ]);
 
   /**
@@ -1049,10 +1052,7 @@ export function Combobox<M extends boolean>({
         }
 
         case keyMap.Enter: {
-          if (!isOpen) {
-            // If the menu is not open, enter should open the menu
-            openMenu();
-          } else if (
+          if (
             // Select the highlighted option iff
             // the menu is open,
             // we're focused on input element,
@@ -1137,6 +1137,7 @@ export function Combobox<M extends boolean>({
    *
    */
 
+  // TODO: Replace this with `useBackdropClick`
   /**
    * We add two event handlers to the document to handle the backdrop click behavior.
    * Intended behavior is to close the menu, and keep focus on the Combobox.