+++ packages/lib/src/index.ts
@@ -5,6 +5,8 @@ import getNodeTextContent from './getNodeTextContent';
 import DarkModeProps, { Theme } from './DarkModeProps';
 import getTheme from './getTheme';
 import { allEqual } from './allEqual';
+export { validateChildren } from './validateChildren';
+export { createSyntheticEvent } from './createSyntheticEvent';
 
 export {
   storybookArgTypes,