+++ packages/search-input/src/SearchResultsMenu/index.ts
@@ -1,0 +1,2 @@
+export { SearchResultsMenu } from './SearchResultsMenu';
+export { SearchResultsMenuProps } from './SearchResultsMenu.types';