+++ packages/segmented-control/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/segmented-control
 
+## 7.0.3
+
+### Patch Changes
+
+- bf2fedf6d: Version bumps lib
+- Updated dependencies [ffb99f417]
+- Updated dependencies [bf2fedf6d]
+- Updated dependencies [b7a29ea38]
+  - @leafygreen-ui/hooks@7.5.0
+  - @leafygreen-ui/leafygreen-provider@3.1.1
+  - @leafygreen-ui/typography@16.0.1
+
 ## 7.0.2
 
 ### Patch Changes